SVG support for django-filer (#1226)
* drop support for high resolution images

* migrate image model fields width and height to float

* SVG images: store width and height in table

* fix #1216: mime-type when using dropzone

* drop quality and other parameters for PNG and SVG in thumb-filename

* fix #1216: mime-type when using dropzone

* Fix: SVG images don’t have an exif header

* Fix: SVG images don’t have an exif header

* do not check for file extension if MIME type is 'application/octet-stream'

* replace all occurences of force_text by force_str

* upgrade to work with nodejs version 14

* move towards templatetag thumbnail

* upgrade to work with nodejs version 14

* proceed with latest versions of npm and gulp

* downgrade phantom and request

* install gulp globally

* add gulp-util to dependencies

* pin nodejs to 14.15.0

* more tests in gulp ci

* try with older versions

* with these versions of node packages, CI runs locally

* remove hacks required for retina.js’s highres images

* introduce templatetag file_icon

* fix some exceptions cause by bad filetype guessing

* Add method ‘exists()’ on MultiStorageFieldFile

* increase icon size in directory listing from 25 to 40

* use SVG icons

* remove png icons and add deprecation warning if used

* add properties mime_maintype and mime_subtype

* make audio element behave like image

* add video element

* render file icon for PDF

* Do not create file icons upon upload

* remove focal-point.js from views without preview image

* remove unused icon

* handle drag & drop of images onto existing plugins

* adopt drag-zone to preview images in 80x80

* adopt drag-zone to preview images in 80x80

* create prerelease of upcoming version of django-filer

* fix scaling of very large images: ceil height to integer

* use exists-method to check for file on disk

* Pin to pre-release of easy-thumbnails (with SVG support)

* fix typo

* fix: add mime_type to filer’s Image rather than Django’s File model

* Fields _width and _height are floats now

* round width and height to one decimal unit

* Drop support for Python-3.5

it is not supported by reportlabs and end of life anyway

* fix E125 continuation line with same indent as next logical line

* Fields _width and _height are floats now

* fix sorting order reproted by isort

* isort complains about one missing empty line

* Bump to version 2.1rc1

* Bump to version 2.1rc1

* simplify code for property exif

* fix #1227: Dropzone layout and render field label

+++ tests/utils/extended_app/migrations/0001_initial.py
@@ -16,8 +16,8 @@ class Migration(migrations.Migration):
         migrations.CreateModel(
             name='ExtImage',
             fields=[
-                ('_height', models.IntegerField(blank=True, null=True)),
-                ('_width', models.IntegerField(blank=True, null=True)),
+                ('_height', models.FloatField(blank=True, null=True)),
+                ('_width', models.FloatField(blank=True, null=True)),
                 ('default_alt_text', models.CharField(blank=True, max_length=255, null=True, verbose_name='default alt text')),
                 ('default_caption', models.CharField(blank=True, max_length=255, null=True, verbose_name='default caption')),
                 ('subject_location', models.CharField(blank=True, default='', max_length=64, verbose_name='subject location')),
